backend/pyproject.toml
[project]
name = "trackai"
version = "0.1.0"
description = "Lightweight experiment tracker for deep learning"
readme = "README.md"
authors = [
    { name = "Túlio Souza", email = "tulio.souza99@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0.0",
    "pyarrow>=18.0.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "duckdb>=1.0.0",
    "duckdb-engine>=0.13.0",
    "click>=8.0.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
trackai = "trackai.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
